add gitignore function
This will pull a useful .gitignore from github/gitignore. The type
to look for is pulled from a command line arg. If no type is
supplied, then an empty .gitignore is created.

NOTE: I wish there was some more exception handling that could be
done here (in the case the user supplies a type that does not match
a type in the github/gitignore repo, or does not spell it correctly).
This may be a good opportunity to switch to something else like
Python, in which case Jinja could be utilized for the README.